+++ src/settings/ModuleSettings.ts
@@ -1,4 +1,4 @@
-import { getGame, localize } from '@/utils/game';
+import { localize } from '@/utils/game';
 import moduleJson from '@module';
 
 export enum SettingKeys {
@@ -13,46 +13,34 @@ type SettingType<K extends SettingKeys> =
     K extends SettingKeys.includedCompendia ? string :
     never;  
 
-// the solo instance
-export let moduleSettings: ModuleSettings;
-
-// set the main application; should only be called once
-export function updateModuleSettings(settings: ModuleSettings): void {
-  moduleSettings = settings;
-}
-
 export class ModuleSettings {
-  constructor() {
-    this.registerSettings();
-  }
-
-  public isSettingValueEmpty(setting: any): boolean {
+  public static isSettingValueEmpty(setting: any): boolean {
     return Object.keys(setting).length === 0 || setting === null || setting === undefined;
   }
 
-  public get<T extends SettingKeys>(setting: T): SettingType<T> {
-    return getGame().settings.get(moduleJson.id, setting) as SettingType<T>;
+  public static get<T extends SettingKeys>(setting: T): SettingType<T> {
+    return game.settings.get(moduleJson.id, setting) as SettingType<T>;
   }
 
-  public async set<T extends SettingKeys>(setting: T, value: SettingType<T>): Promise<void> {
-    await getGame().settings.set(moduleJson.id, setting, value);
+  public static async set<T extends SettingKeys>(setting: T, value: SettingType<T>): Promise<void> {
+    await game.settings.set(moduleJson.id, setting, value);
   }
 
-  private register(settingKey: string, settingConfig: ClientSettings.PartialSettingConfig) {
-    getGame().settings.register(moduleJson.id, settingKey, settingConfig);
+  private static register(settingKey: string, settingConfig: ClientSettings.PartialSettingConfig) {
+    game.settings.register(moduleJson.id, settingKey, settingConfig);
   }
 
-  private registerMenu(settingKey: string, settingConfig: ClientSettings.PartialSettingSubmenuConfig) {
-    getGame().settings.registerMenu(moduleJson.id, settingKey, settingConfig);
+  private static registerMenu(settingKey: string, settingConfig: ClientSettings.PartialSettingSubmenuConfig) {
+    game.settings.registerMenu(moduleJson.id, settingKey, settingConfig);
   }
 
   // these are local menus (shown at top)
-  private localMenuParams: (ClientSettings.PartialSettingSubmenuConfig & { settingID: string })[] = [
+  private static localMenuParams: (ClientSettings.PartialSettingSubmenuConfig & { settingID: string })[] = [
   ];
 
   // these are globals shown in the options
   // name and hint should be the id of a localization string
-  private displayParams: (ClientSettings.PartialSettingConfig & { settingID: string })[] = [
+  private static displayParams: (ClientSettings.PartialSettingConfig & { settingID: string })[] = [
     {
       settingID: SettingKeys.resultLength,
       name: 'acm.settings.resultLength',
@@ -70,21 +58,21 @@ export class ModuleSettings {
   ];
 
   // these are client-specific and displayed in settings
-  private localDisplayParams: (ClientSettings.PartialSettingConfig & { settingID: string })[] = [
+  private static localDisplayParams: (ClientSettings.PartialSettingConfig & { settingID: string })[] = [
   ];
 
   // these are globals only used internally
-  private internalParams: (ClientSettings.PartialSettingConfig & { settingID: string })[] = [
+  private static internalParams: (ClientSettings.PartialSettingConfig & { settingID: string })[] = [
   ];
   
   // these are client-specfic only used internally
-  private localInternalParams: (ClientSettings.PartialSettingConfig & { settingID: string })[] = [
+  private static localInternalParams: (ClientSettings.PartialSettingConfig & { settingID: string })[] = [
   ];
 
-  private registerSettings(): void {
-    for (let i=0; i<this.localMenuParams.length; i++) {
-      const { settingID, ...settings} = this.localMenuParams[i];
-      this.registerMenu(settingID, {
+  public static registerSettings(): void {
+    for (let i=0; i<ModuleSettings.localMenuParams.length; i++) {
+      const { settingID, ...settings} = ModuleSettings.localMenuParams[i];
+      ModuleSettings.registerMenu(settingID, {
         ...settings,
         name: settings.name ? localize(settings.name) : '',
         hint: settings.hint ? localize(settings.hint) : '',
@@ -92,9 +80,9 @@ export class ModuleSettings {
       });
     }
 
-    for (let i=0; i<this.displayParams.length; i++) {
-      const { settingID, ...settings} = this.displayParams[i];
-      this.register(settingID, {
+    for (let i=0; i<ModuleSettings.displayParams.length; i++) {
+      const { settingID, ...settings} = ModuleSettings.displayParams[i];
+      ModuleSettings.register(settingID, {
         ...settings,
         name: settings.name ? localize(settings.name) : '',
         hint: settings.hint ? localize(settings.hint) : '',
@@ -103,9 +91,9 @@ export class ModuleSettings {
       });
     }
 
-    for (let i=0; i<this.localDisplayParams.length; i++) {
-      const { settingID, ...settings} = this.localDisplayParams[i];
-      this.register(settingID, {
+    for (let i=0; i<ModuleSettings.localDisplayParams.length; i++) {
+      const { settingID, ...settings} = ModuleSettings.localDisplayParams[i];
+      ModuleSettings.register(settingID, {
         ...settings,
         name: settings.name ? localize(settings.name) : '',
         hint: settings.hint ? localize(settings.hint) : '',
@@ -114,18 +102,18 @@ export class ModuleSettings {
       });
     }
 
-    for (let i=0; i<this.internalParams.length; i++) {
-      const { settingID, ...settings} = this.internalParams[i];
-      this.register(settingID, {
+    for (let i=0; i<ModuleSettings.internalParams.length; i++) {
+      const { settingID, ...settings} = ModuleSettings.internalParams[i];
+      ModuleSettings.register(settingID, {
         ...settings,
         scope: 'world',
         config: false,
       });
     }
 
-    for (let i=0; i<this.localInternalParams.length; i++) {
-      const { settingID, ...settings} = this.localInternalParams[i];
-      this.register(settingID, {
+    for (let i=0; i<ModuleSettings.localInternalParams.length; i++) {
+      const { settingID, ...settings} = ModuleSettings.localInternalParams[i];
+      ModuleSettings.register(settingID, {
         ...settings,
         scope: 'client',
         config: false,